requirements.txt
# Common Python dependencies for BloomFlow services
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
python-keycloak==2.10.0

